refactor: refactored the testset table component code for better readability - separated sub-components into new components - renamed the loading state hook - aligned all the codes - improved function logic - created a new function named createNewColDefs to avoid code duplication
